\name{get_wdgt_vals}
\alias{get_wdgt_vals}
\title{Get Widget Values - Internal Function}
\usage{
get_wdgt_vals(widget_list)
}
\arguments{
  \item{widget_list}{A list of widgets}
}
\description{
The \code{get_wdgt_vals} is the internal function that
implements the routine to get the current value from all
the widgets in the list.
}
\details{
This function, with a widgets list, get the current value
from all the widgets in the list as a vector of characters.
}
\seealso{
\code{\link{set_wdgt_vals}}; \code{\link{set_wdgt_sel}}
}

