% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_dcf_ind.R
\name{gui_dcf_ind}
\alias{gui_dcf_ind}
\title{DCF indicators GUI}
\usage{
gui_dcf_ind()
}
\value{
This function does not return a value.
}
\description{
The \code{gui_dcf_ind} function implements the graphical user interface for the
 DCF indicators computation routine.
}
\details{
This function,  with a count vector generated by \code{\link{gui_out_grid}},
 computes the DCF indicators 5 and 6.
}
\references{
Russo, T., Parisi, A., & Cataudella, S. (2013). Spatial indicators of fishing pressure: Preliminary analyses and possible developments. \emph{Ecological indicators}, \bold{26}, 141--153.
\url{http://www.sciencedirect.com/science/article/pii/S1470160X12003822}
}
\seealso{
\code{\link{gui_out_grid}}
}
