\name{gui_join_lb_vms}
\alias{gui_join_lb_vms}
\title{VMS-LB Match GUI}
\usage{
gui_join_lb_vms(lb_db_name = "", vms_db_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}

  \item{lb_db_name}{The path of a LogBook DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_join_lb_vms} function implements the
graphical user interface for the VMS-LB Matching routine.
}
\details{
This function, with both VMS and LogBook databases,
performs a VMS-LB track match based on vessel, track and
time conformity.
}
\references{
free text reference Pointers to the literature related to
this object.
}

