% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_lb_db_create.R
\name{gui_lb_db_create}
\alias{gui_lb_db_create}
\title{LogBook DataBase Creation GUI}
\usage{
gui_lb_db_create()
}
\value{
This function does not return a value. 
After the execution a logbook database will be deployed.
}
\description{
The \code{gui_lb_db_create} function implements the graphical user interface for the
creation of a logbook database.
This function, with an edited logbook dataset (see \code{\link{gui_lb_editraw}}),
 creates a new logbook database.
}
\seealso{
\code{\link{gui_lb_editraw}}
}

