% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_lb_db_edit.R
\name{gui_lb_db_edit}
\alias{gui_lb_db_edit}
\title{Logbook Editing GUI}
\usage{
gui_lb_db_edit(lb_db_name = "")
}
\arguments{
\item{lb_db_name}{The path of a Logbook DataBase}
}
\value{
This function does not return a value. 
After the execution, the Logbook DB will be updated with the edited format.
}
\description{
The \code{gui_lb_db_edit} function implements the graphical user interface for the
 editing of a LogBook Database
}
\details{
In this gui, with a LogBook Database, the user can automatically edit the logbook
DB.
}
