% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_main.R
\name{gui_main}
\alias{gui_main}
\title{VMSbase main GUI}
\usage{
gui_main()
}
\value{
This function does not return a value.
}
\description{
The \code{gui_main} function implements the main graphical user interface of the
 VMSbase package
}
\details{
With this gui the user can access all the functionalities of VMSbase
}

