\name{gui_mark_fis_poi}
\alias{gui_mark_fis_poi}
\title{Mark Fishing Points GUI}
\usage{
gui_mark_fis_poi(vms_db_name = "", harb_file_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}

  \item{harb_file_name}{The path of a shape file with
  harbours point data}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_mark_fis_poi} function implements the
graphical user interface for the Fishing Point Marking
routine.
}
\details{
This function, with a VMS database and a shape file with
harbours points, performs a filtered search over the whole
db assigning fishing status to the vms interpolated data.
}
\references{
free text reference Pointers to the literature related to
this object.
}

