% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_db_are.R
\name{gui_vms_db_are}
\alias{gui_vms_db_are}
\title{VMS DB Area Assignment GUI}
\usage{
gui_vms_db_are(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_are} function implements the graphical user interface for the
 VMS Assign Area routine.
}
\details{
This function, with a VMS database,
 assigns areas to VMS tracks according to median positions.
}
