\name{gui_vms_db_create}
\alias{gui_vms_db_create}
\title{VMS DataBase Creation GUI}
\usage{
gui_vms_db_create()
}
\value{
This function does not return a value. After the execution
a VMS DataBase will be deployed.
}
\description{
The \code{gui_vms_db_create} function implements the
graphical user interface for the creation of a VMS
DataBase.
}
\details{
This function, with an edited vms dataset (see
\code{\link{gui_vms_editraw}}), creates a new VMS DataBase.
}
\references{
free text reference Pointers to the literature related to
this object.
}
\seealso{
\code{\link{gui_vms_editraw}}
}

