% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_db_cut.R
\name{gui_vms_db_cut}
\alias{gui_vms_db_cut}
\title{VMS DB Cut GUI
 
The \code{gui_vms_db_cut} function implements the graphical user interface for the
 VMS Track cutting routine.}
\usage{
gui_vms_db_cut(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
This function, with a VMS cleaned database (see \code{\link{gui_vms_db_clean}}),
 assign track numbers to the raw pings of each vessel in the VMS DB.
}
\seealso{
\code{\link{gui_vms_db_clean}}
}

