\name{gui_vms_db_dep}
\alias{gui_vms_db_dep}
\title{VMS DB Depth Assignment GUI}
\usage{
gui_vms_db_dep(vms_db_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_dep} function implements the graphical
user interface for the VMS Assign Depth routine.
}
\details{
This function, with a VMS database, assigns depths to VMS
interpolated points according to positions on the 3D spline
of XYZ bathymetry data downloaded from NOAA servers.
}
\references{
free text reference Pointers to the literature related to
this object.
}

