\name{gui_vms_db_intr}
\alias{gui_vms_db_intr}
\title{VMS DB Interpolation GUI}
\usage{
gui_vms_db_intr(vms_db_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_intr} function implements the
graphical user interface for the VMS Track Interpolation
routine.
}
\details{
This function, with a VMS cleaned and cutted database (see
\code{\link{gui_vms_db_cut}} and
\code{\link{gui_vms_db_clean}}), interpolates track data of
each vessel in the VMS DB.
}
\references{
free text reference Pointers to the literature related to
this object.
}
\seealso{
\code{\link{gui_vms_db_cut}} \code{\link{gui_vms_db_clean}}
}

