\name{gui_vms_db_stat}
\alias{gui_vms_db_stat}
\title{VMS DataBase Status GUI}
\usage{
gui_vms_db_stat(vms_db_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_stat} function implements the
graphical user interface for the VMS DataBase Status
viewer.
}
\details{
This function, with a VMS DataBase, shows the current VMS
DataBase status.
}
\references{
free text reference Pointers to the literature related to
this object.
}

