% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_save_bat.R
\name{gui_vms_save_bat}
\alias{gui_vms_save_bat}
\title{VMS DB Save Bathymetry GUI
 
The \code{gui_vms_save_bat} function implements the graphical user interface for the
 VMS Save Bathymetry routine.}
\usage{
gui_vms_save_bat(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
This function, with a VMS database,
 downloads an XYZ bathymetry file from NOAA servers.
}
\seealso{
\code{\link{gui_vms_view_ping}} \code{\link{gui_vms_view_track}} \code{\link{gui_vms_view_intrp}}
}
