% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_view_intrp.R
\name{gui_vms_view_intrp}
\alias{gui_vms_view_intrp}
\title{VMS DB View Interpolated Data GUI}
\usage{
gui_vms_view_intrp(vms_db_name = "", bathy_file_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}

\item{bathy_file_name}{The path of a Bathymetry file}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_view_intrp} function implements the graphical user interface for the
 VMS DB routine to view interpolated data.
}
\details{
This function, with a VMS database,
 plots vessel and track data.
}
\seealso{
\code{\link{gui_vms_view_ping}} \code{\link{gui_vms_view_track}} \code{\link{gui_vms_save_bat}}
}
