% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_view_track.R
\name{gui_vms_view_track}
\alias{gui_vms_view_track}
\title{VMS DB View Track Data GUI
 
The \code{gui_vms_view_track} function implements the graphical user interface for the
 VMS DB routine to view track data.}
\usage{
gui_vms_view_track(vms_db_name = "", bathy_file_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}

\item{bathy_file_name}{The path of a Bathymetry file}
}
\value{
This function does not return a value.
}
\description{
This function, with a VMS database,
 plots vessel and track data.
}
\seealso{
\code{\link{gui_vms_view_ping}} \code{\link{gui_vms_save_bat}} \code{\link{gui_vms_view_intrp}}
}
