% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_viz_adv.R
\name{gui_vms_viz_adv}
\alias{gui_vms_viz_adv}
\title{VMS DB View Track Data GUI
 
The \code{gui_vms_view_track} function implements the graphical user interface for the
 VMS DB routine to view vessel and track data with the Google Maps capabilities.}
\usage{
gui_vms_viz_adv(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
This function, with a VMS database,
 plots vessel and track data on Hybrid Satellite images from Google Maps.
}
\seealso{
\code{\link{gui_vms_view_ping}} \code{\link{gui_vms_view_track}} \code{\link{gui_vms_view_intrp}}
}

