% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_mixsou.R
\name{gui_vmsdb_mixsou}
\alias{gui_vmsdb_mixsou}
\title{VMS DataBase Creation GUI}
\usage{
gui_vmsdb_mixsou()
}
\value{
This function does not return a value.
After the execution a VMS DataBase will be deployed.
}
\description{
The \code{gui_vmsdb_mixsou} function implements the graphical user interface for the
creation of a VMS DataBase from multiple sources.
}
\details{
This function, with an edited vms dataset (see \code{\link{gui_vms_editraw}}),
creates a new VMS DataBase.
}
\seealso{
\code{\link{gui_vms_editraw}}
}
