% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{harbCoo}
\alias{harbCoo}
\alias{harbCoo-class}
\title{Harbours Coordinates Shape File class}
\description{
Harbours coordinates Shape File reference Class

\itemize{
\item{\code{path} character The Path to the harbours coordinates shape file}
\item{\code{data} SpatialPointsDataFrame The loaded harbours coordinates data}

}
}
\details{
\code{harbCoo} class is a reference class for the Harbours Coordinates shape file.
}

