\name{latsex2dec}
\alias{latsex2dec}
\title{Sexagesimal to Decimal Latitude conversion function}
\usage{
latsex2dec(degree, minute, second, direction)
}
\arguments{
  \item{degree}{The value in degrees of the latitude.}

  \item{minute}{The value in minutes of the latitude.}

  \item{second}{The value in seconds of the latitude.}

  \item{direction}{The direction, as "N" or "S", of the
  latitude.}
}
\value{
The function returns the latitude value converted in the
decimal numeral system
}
\description{
\code{latsex2dec} converts sexagesimal latitude data to
decimal latitude values.
}

