\docType{class}
\name{log_File}
\alias{log_File}
\alias{log_File-class}
\title{LogBook File Class}
\description{
LogBook File reference Class

\itemize{ \item{\code{path} character path of the logbook
file} \item{\code{data} data.frame data from the logbook
file}

}
}
\details{
\code{log_File} class is a reference class for the LogBook
file.
}

