% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_vms_db_clean.R
\name{plotNet}
\alias{plotNet}
\title{Neural Network Plotting - Internal Function}
\usage{
plotNet(net, Dg, ConfMat)
}
\arguments{
\item{net}{The Neural Network}

\item{Dg}{The Prediction Index}

\item{ConfMat}{The Confusion Matrix}
}
\value{
This function does not return a value.
}
\description{
The \code{plotNet} is the internal function that implements the
 Neural Network Plotting routine.
}
\details{
This function,  with a Neural Network, a Prediction Index and a Confusion Matrix, generated by \code{\link{gui_vms_met_pred}},
 plots the graphical overall result of the neural network prediction.
}
\references{
free text reference Pointers to the literature related to this object.
}
\seealso{
\code{\link{gui_vms_met_pred}}
}
