% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{polymap}
\alias{polymap}
\alias{polymap-class}
\title{Land Map Shape File class}
\arguments{
\item{path}{character The Path to the Land Map shape file}

\item{data}{SpatialPolygonsDataFrame The loaded Land Map data}
}
\description{
Land Map Shape File reference Class

\itemize{
\item{\code{path} character The Path to the Land Map shape file}
\item{\code{data} SpatialPolygonsDataFrame The loaded Land Map data}

}
}
\details{
\code{polymap} class is a reference class for the Land Map shape file.
}

