% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{que_vms_DB}
\alias{que_vms_DB}
\alias{que_vms_DB-class}
\title{VMS DataBase Query class}
\description{
VMS DataBase Query reference Class

\itemize{
\item{\code{dir} character path of the VMS DataBase}
\item{\code{db} character name of the VMS DataBase}
\item{\code{que} character query for the VMS DataBase}

}
}
\details{
\code{que_vms_DB} class is a reference class for the VMS DataBase.
}

