% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to Degree Angles conversion function}
\usage{
rad2deg(heading)
}
\arguments{
\item{heading}{The value in radians of the heading.}
}
\value{
The function returns the heading value converted in degrees
}
\description{
\code{rad2deg} converts radian heading data to degree values.
}

