% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRawVms.R
\name{saveRawVms}
\alias{saveRawVms}
\title{Raw VMS editing function}
\usage{
saveRawVms(rawfile, widget)
}
\arguments{
\item{rawfile}{The raw VMS dataset.}

\item{widget}{The widget list that contains the editing infos.}
}
\value{
The function returns the standardized VMS data.
}
\description{
\code{saveRawVms} implements the routines that converts raw values
 to standard data.
}
\seealso{
\code{\link{gui_vms_editraw}}
}
