\name{turn_wdgt_off}
\alias{turn_wdgt_off}
\title{Turn Widget Off - Internal Function}
\usage{
turn_wdgt_off(widget_list)
}
\arguments{
  \item{widget_list}{A list of widgets}
}
\description{
The \code{turn_wdgt_off} is the internal function that
implements the routine to switch off all the widgets in the
list.
}
\details{
This function, with a widgets list, turns off all the
widgets in the list.
}
\seealso{
\code{\link{turn_wdgt_on}}
}

