\name{turn_wdgt_on}
\alias{turn_wdgt_on}
\title{Turn Widget On - Internal Function}
\usage{
turn_wdgt_on(widget_list)
}
\arguments{
  \item{widget_list}{A list of widgets}
}
\description{
The \code{turn_wdgt_on} is the internal function that
implements the routine to switch on all the widgets in the
list.
}
\details{
This function, with a widgets list, turns on all the
widgets in the list.
}
\seealso{
\code{\link{turn_wdgt_off}}
}

