% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{vms_File}
\alias{vms_File}
\alias{vms_File-class}
\title{VMS File Class}
\description{
VMS File reference Class
\itemize{
\item{\code{path} character path of the vms file}
\item{\code{data} data.frame data from the vms file}
}
}
\details{
\code{vms_File} class is a reference class for the VMS file.
}

