\name{gui_dcf_ind}
\alias{gui_dcf_ind}
\title{DCF indicators GUI}
\usage{
gui_dcf_ind()
}
\value{
This function does not return a value.
}
\description{
The \code{gui_dcf_ind} function implements the graphical
user interface for the DCF indicators computation routine.
}
\details{
This function, with a count vector generated by
\code{\link{gui_out_grid}}, computes the DCF indicators 5
and 6.
}
\references{
free text reference Pointers to the literature related to
this object.
}
\seealso{
\code{\link{gui_out_grid}}
}

