% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_lb_met_edi.R
\name{gui_lb_met_edi}
\alias{gui_lb_met_edi}
\title{Metier Editing GUI}
\usage{
gui_lb_met_edi()
}
\value{
This function does not return a value. 
After the execution, the edited names will be added to the LogBook Metier file.
}
\description{
The \code{gui_lb_met_edi} function implements the graphical user interface for the
 editing of a LogBook Metier file
}
\details{
In this gui, with a LogBook Metier file, the user can change the simple discovery
 ordering of the clusters to a corresponding international metier code.
}
