\name{gui_vms_db_cut}
\alias{gui_vms_db_cut}
\title{VMS DB Cut GUI}
\usage{
gui_vms_db_cut(vms_db_name = "")
}
\arguments{
  \item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_cut} function implements the graphical
user interface for the VMS Track cutting routine.
}
\details{
This function, with a VMS cleaned database (see
\code{\link{gui_vms_db_clean}}), assign track numbers to
the raw pings of each vessel in the VMS DB.
}
\references{
free text reference Pointers to the literature related to
this object.
}
\seealso{
\code{\link{gui_vms_db_clean}}
}

