% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_db_intr.R
\name{gui_vms_db_intr}
\alias{gui_vms_db_intr}
\title{VMS DB Interpolation GUI}
\usage{
gui_vms_db_intr(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
The \code{gui_vms_db_intr} function implements the graphical user interface for the
 VMS Track Interpolation routine.
}
\details{
This function, with a VMS cleaned and cutted database (see \code{\link{gui_vms_db_cut}} and \code{\link{gui_vms_db_clean}}),
 interpolates track data of each vessel in the VMS DB.
}
\references{
Russo, T., Parisi, A. and Cataudella, S. (2011) New insights in interpolating fishing tracks from VMS data for different metiers. \emph{Fisheries Research}, \bold{108(1)}, 184--194.
\url{http://www.sciencedirect.com/science/article/pii/S0165783610003450}
}
\seealso{
\code{\link{gui_vms_db_cut}} \code{\link{gui_vms_db_clean}}
}
