% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_db_select.R
\name{gui_vms_db_sel}
\alias{gui_vms_db_sel}
\title{VMS DataBase Select GUI
 
The \code{gui_vms_db_sel} function implement the graphical user interface for the
 VMS DataBase Select routine.}
\usage{
gui_vms_db_sel(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
This function, with a VMS DataBase (see \code{\link{gui_vms_db_stat}}),
 enables the user to perform queries on, and extract data from, the submitted VMS DataBase.
}
\seealso{
\code{\link{gui_vms_db_stat}}
}
