% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_vms_met_pred.R
\name{gui_vms_met_pred}
\alias{gui_vms_met_pred}
\title{Predict Metier GUI
 
The \code{gui_vms_met_pred} function implements the graphical user interface for the
 Metier Prediction}
\usage{
gui_vms_met_pred(vms_db_name = "")
}
\arguments{
\item{vms_db_name}{The path of a VMS DataBase}
}
\value{
This function does not return a value.
}
\description{
This function,  with a VMS database and a shape file with harbours points, performs a neural network prediction over the whole
 db assigning metier data to vms tracks based on a training with existing vms-lb match data
  given by the database.
}
\references{
Russo, T., Parisi, A., Prorgi, M., Boccoli, F., Cignini, I., Tordoni, M. and Cataudella, S. (2011) When behaviour reveals activity: Assigning fishing effort to metiers based on VMS data using artificial neural networks. \emph{Fisheries Research}, \bold{111(1)}, 53--64.
\url{http://www.sciencedirect.com/science/article/pii/S0165783611002281}
}

