% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{log_Cla}
\alias{log_Cla}
\alias{log_Cla-class}
\title{LogBook Clustering class}
\description{
LogBook Clustering rData file reference Class

\itemize{
\item{\code{data} character The Clustering result data}
\item{\code{options} character Metier manual annotation}

}
}
\details{
\code{log_Cla} class is a reference class for the Logbook Clustering rData file.
}

