% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonsex2dec.R
\name{lonsex2dec}
\alias{lonsex2dec}
\title{Sexagesimal to Decimal Longitude conversion function}
\usage{
lonsex2dec(degree, minute, second, direction)
}
\arguments{
\item{degree}{The value in degrees of the longitude.}

\item{minute}{The value in minutes of the longitude.}

\item{second}{The value in seconds of the longitude.}

\item{direction}{The direction, as "E" or "W", of the longitude.}
}
\value{
The function returns the longitude value converted in the decimal numeral system
}
\description{
\code{lonsex2dec} converts sexagesimal longitude data to decimal longitude values.
}

