% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRawLogBook.R
\name{saveRawLogBook}
\alias{saveRawLogBook}
\title{Raw LogBook editing function}
\usage{
saveRawLogBook(rawfile, widget)
}
\arguments{
\item{rawfile}{The raw LogBook dataset.}

\item{widget}{The widget list that contains the editing infos.}
}
\value{
The function returns the standardized VMS data.
}
\description{
\code{saveRawLogBook} implements the routines that converts raw values
 to standard data.
}
\seealso{
\code{\link{gui_lb_editraw}}
}
